add example callbacks to Node for status and 'requests'
included some placeholder examples from the sample I used in building
the addon.  The sample had provisions for invoking a callback.  That
knowledge is captured in addonTest3 in the form of a get_status and
send_request method - both of these will invoke a callback when
completed.  Ultimately a callback would be used to return any data
queried from the teensy.